TSM-135 - discuss an event, *fixed build/jar problem

+++ trunk/tsm/src/test/java/com/tech4d/tsm/audit/IntegrationTestAuditEntry.java
@@ -209,7 +209,7 @@ public class IntegrationTestAuditEntry extends OpenSessionInViewIntegrationTest 
     	assertEquals(3, auditEntryDao.getAuditEntries(USERNAME, Event.class, 0, 20).size());
     	assertEquals(3L, (long)auditEntryDao.getAuditEntriesCount(USERNAME, Event.class));
     	assertEquals(5, auditEntryDao.getAuditEntries(USERNAME2, Event.class, 0, 20).size());
-    	assertEquals(5L, auditEntryDao.getAuditEntriesCount(USERNAME2, Event.class));
+    	assertEquals(5L, (long)auditEntryDao.getAuditEntriesCount(USERNAME2, Event.class));
     	//limit size of results
     	List<AuditUserChange> entries = auditEntryDao.getAuditEntries(USERNAME2, Event.class, 0, 3);
     	AuditUserChange first = entries.get(0);
